requirements.txt
Django==2.2
django-environ==0.4.5
django-admin-tools==0.8.1
djangorestframework==3.10.3
django-multiselectfield==0.1.11
mysqlclient==1.4.6
numpy==1.17.0
pandas==0.25.0
redis==3.3.2
